track_separation/requirements.txt
absl-py==2.1.0
anyio==3.7.1
asgiref==3.8.1
astunparse==1.6.3
async-timeout==5.0.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==7.1.2
Django==4.2.16
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
flatbuffers==1.12
future==1.0.0
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.66.2
h11==0.12.0
h2==4.1.0
h5py==3.12.1
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.0.1
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
norbert==0.2.1
numpy==1.23.5
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.1
pandas==1.5.3
pillow==10.4.0
protobuf==3.19.6
psutil==6.0.0
psycopg2-binary==2.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydub==0.25.1
pyffmpeg==2.4.2.18.1
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
spleeter==2.4.0
sqlparse==0.5.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
typer==0.3.2
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
